new version-with visualisation
I don't know how to edit directly in github, since it shows "invalid" after I edit it. So I just upload a new version for it. And the plot of CLT and bootstrapping seems totally the same, so I think there's might some errors, please feel free to point it out and I will fix it later.